feat: testes automatizados comc ypress

+++ src/services/Short.service.ts
@@ -1,7 +1,6 @@
 import moment from "moment";
 import { ShortResponse } from "../interfaces/ShortResponse";
 import StorageService from "./Storage.service";
-import ToastService from "./Toast.service";
 
 class ShortService {
   private url = "https://url.api.stdlib.com/temporary@0.3.0/";
@@ -15,27 +14,31 @@ class ShortService {
       method: "POST",
       body: JSON.stringify(input),
     });
-    const link = (await response.json()) as ShortResponse;
+    const body = response;
+    if (body.status !== 200) {
+      throw new Error("Falha ao encurtar link. Tente novamente.");
+    }
+    const link = (await body.json()) as ShortResponse;
     StorageService.saveLinks({ ...link, dateExpires: this.getDateExpires() });
     return link;
   }
 
   async destroyLink(input: { key: string }) {
-    try {
-      const response = await fetch(this.url + "destroy", {
-        method: "POST",
-        body: JSON.stringify(input),
-      });
-      const data = (await response.json()) as boolean;
-      if (!data) {
-        throw new Error(
-          "Não foi possível excluir o link. Tente novamente mais tarde."
-        );
-      }
-      StorageService.removeLink(input.key);
-    } catch (error: any) {
-      ToastService.toastError({ message: error.message });
+    const response = await fetch(this.url + "destroy", {
+      method: "POST",
+      body: JSON.stringify(input),
+    });
+    const body = response;
+    if (body.status !== 200) {
+      throw new Error("Falha ao remover link. Tente novamente.");
     }
+
+    const data = (await response.json()) as boolean;
+    if (!data) {
+      throw new Error("Falha ao remover link. Tente novamente.");
+    }
+
+    StorageService.removeLink(input.key);
   }
 
   private getDateExpires(): string {
